Button press with sound
I created a button press with a cheesy way to get sound to play once at a time.  It is by no means the correct way to do it, but it is a workaround. Became more familiar with the logic of blueprints through a math function and some input relations.