Button to Fan Rotation
Successful blueprint communication. I was able to get the fan to rotate when the button was pressed. I also got a cone to rotate, but it does not rotate around the origin of the fan. I will need to either set the origin of every object I want to rotate in the editor, or set a point to rotate around in the blueprint classes.